MeshArrays.jl documentation

MeshArrays.jl documentation

MeshArrays.jl primarily defines composite types that embed inter-connected array collections within a struct and provides an exchange function that effectively transfers data between connected arrays. It was originally introduced, as gcmfaces.jl, in this JuliaCon-2018 presentation (see below for notebooks). Note: MeshArrays.jl is registered, documented, archived, and routinely tested, but is still regarded as a preliminary implementation.

Contents

MeshArrays.jl documentation

Contents

Installation

Use examples

Main Features

API index

API details

Installation

using Pkg
Pkg.add("MeshArrays")
Pkg.test("MeshArrays")

Note: Julia's package manager is documented here within docs.julialang.org.

Use examples

The JuliaCon-2018 presentation relied on two Jupyter notebooks that are available in the JuliaCon2018Notebooks repo and pre-defined grids that are available from this website and can e.g. be downloaded as follows:

setenv DemoGrids 'ftp://mit.ecco-group.org/gforget/grid_examples/'
wget --recursive {$DemoGrids}/GRID_CS32/
wget --recursive {$DemoGrids}/GRID_LLC90/
mv mit.ecco-group.org/gforget/grid_examples/GRID_* .

After downloading GRID_LLC90/, you can also replicate this example using the following commands:

using MeshArrays

!isdir("GRID_LLC90") ? error("could not find GRID_LLC90/") : nothing
(D,Dexch,Darr,DD)=demo1("LLC90")
(Rini,Rend,DXCsm,DYCsm)=demo2()

include(joinpath(dirname(pathof(MeshArrays)),"gcmfaces_plot.jl"))
qwckplot(Rini)
qwckplot(Rend)

Main Features

MeshArrays.jl composite types contain array collections where arrays are typically inter-connected at their edges. It provides exchange functions that transfer data between neighbor arrays so that the user can extend the domain of computation as needed e.g. for interpolation or to compute spatial derivatives.

The composite types specify how each array collection forms a mesh and provide information to allow exchange to dispatch to the appropriate method. Various configurations that are commonly used in Earth System Models are implemented using the concrete type called gcmfaces. Embedded arrays, or meshes, each represent a subdomain within, e.g., an Earth System Model grid.

The gcmfaces name derives from a previous Matlab / Octave package that was introduced in Forget et al., 2015, doi:10.5194/gmd-8-3071-2015, and inspired this Julia package. Here, GCM is an acronym for General Circulation Model, or Global Climate Model as described in this wikipedia entry, and faces is just another name for meshes, arrays, facets, or subdomains.

API index

MeshArrays.gcmfaces

MeshArrays.gcmfaces

MeshArrays.gcmfaces

MeshArrays.gcmfaces

MeshArrays.gcmfaces

MeshArrays.gcmsubset

MeshArrays.GCMGridLoad

MeshArrays.GCMGridOnes

MeshArrays.GCMGridSpec

MeshArrays.LatCircles

MeshArrays.TransportThrough

MeshArrays.convergence

MeshArrays.convert2array

MeshArrays.convert2array

MeshArrays.convert2gcmfaces

MeshArrays.convert2gcmfaces

MeshArrays.demo1

MeshArrays.demo2

MeshArrays.demo3

MeshArrays.exchange

MeshArrays.fijind

MeshArrays.fsize

MeshArrays.fsize

MeshArrays.fsize

MeshArrays.fsize

MeshArrays.gradient

MeshArrays.gradient

MeshArrays.gradient

MeshArrays.mask

MeshArrays.mask

MeshArrays.mask

MeshArrays.read_bin

MeshArrays.read_bin

MeshArrays.read_bin

MeshArrays.smooth

API details

MeshArrays.GCMGridLoad — Method.

GCMGridLoad()

Loads grid variables from files located in grDir set by GCMGridSpec.

Grid variables are XC, XG, YC, YG, RAC, RAZ, DXC, DXG, DYC, DYG, hFacC, hFacS, hFacW, Depth based on the MITgcm naming convention.

source

MeshArrays.GCMGridOnes — Method.

GCMGridOnes(grTp,nF,nP)

Define all-1 grid variables instead of using GCMGridSpec & GCMGridLoad.

source

MeshArrays.GCMGridSpec — Function.

GCMGridSpec(gridName)

Set global variables in the module scope for grDir, nFaces, grTopo, ioSize, facesSize, ioPrec using hard-coded values for LLC90, CS32, LL360 (for now).

source

MeshArrays.convergence — Method.

convergence(uFLD::gcmfaces,vFLD::gcmfaces)

Compute convergence of a vector field

source

MeshArrays.convert2array — Method.

convert2array(fld::gcmfaces)

Convert gcmfaces to array

source

MeshArrays.convert2array — Method.

convert2array(fld::Array{T,N}) where {T,N}

Convert array to gcmfaces

source

MeshArrays.convert2gcmfaces — Method.

convert2gcmfaces(fld::Array)

Convert gcmfaces to mitgcm output

source

MeshArrays.convert2gcmfaces — Method.

convert2gcmfaces(fld::gcmfaces)

Convert mitgcm output to gcmfaces

source

MeshArrays.demo1 — Method.

demo1(gridChoice)

Demonstrate basic fucntions (arithmetic, exchange, GCMGridLoad, gradient, etc.). Example call:

isdir("GRID_LLC90") ? (D,Dexch,Darr,DD)=demo1("LLC90") : nothing

source

MeshArrays.demo2 — Method.

demo2()

Demonstrate higher level functions using smooth() and

isdir("GRID_LLC90") ? demo1("LLC90") : GCMGridOnes("cs",6,100)
(Rini,Rend,DXCsm,DYCsm)=demo2()
@time Rend=smooth(Rini,DXCsm,DYCsm)

include(joinpath(dirname(pathof(MeshArrays)),"gcmfaces_plot.jl"))
qwckplot(Rini)
qwckplot(Rend)

source

MeshArrays.demo3 — Method.

demo3()

Demonstrate computations of ocean meridional transports. Calling sequence:

!isdir("GRID_LLC90")||!isdir("nctiles_climatology") ? error("missing files") : nothing

GCMGridSpec("LLC90")
GCMGridLoad()

include(joinpath(dirname(pathof(MeshArrays)),"gcmfaces_nctiles.jl"))
fileName="nctiles_climatology/UVELMASS/UVELMASS"
U=read_nctiles(fileName,"UVELMASS");
fileName="nctiles_climatology/VVELMASS/VVELMASS"
V=read_nctiles(fileName,"VVELMASS");

(UV, LC, Tr)=demo3(U,V);

using Statistics
include(joinpath(dirname(pathof(MeshArrays)),"gcmfaces_plot.jl"))
qwckplot(UV["U"][:,:,1,1],"U component (note varying face orientations)")
qwckplot(UV["V"][:,:,1,1],"V component (note varying face orientations)")
plot(dropdims(mean(sum(Tr,dims=2),dims=3),dims=(2,3))/1e6,title="meridional transport")

source

MeshArrays.exchange — Method.

exchange(fld::gcmfaces)

Exchange / transfer data between neighbor arrays. Other methods are

exchange(fld::gcmfaces,N::Integer)
exchange(u::gcmfaces,v::gcmfaces)
exchange(u::gcmfaces,v::gcmfaces,N::Integer)

source

MeshArrays.fijind — Method.

fijind(A::gcmfaces,ij::Int)

fijind convenience function.

source

MeshArrays.fsize — Method.

fsize(A::AbstractGcmfaces{T, N}) where {T,N}

fsize convenience function.

source

MeshArrays.fsize — Method.

fsize(A::Array{Array{T,N},1}) where {T,N}

fsize convenience function.

source

MeshArrays.fsize — Method.

fsize(A::AbstractGcmfaces{T, N},i::Int) where {T,N}

fsize convenience function.

source

MeshArrays.fsize — Method.

fsize(A::Array{Array{T,N},1},i::Int) where {T,N}

fsize convenience function.

source

MeshArrays.gradient — Method.

gradient(inFLD::gcmfaces,doDIV::Bool)

Compute spatial derivatives with or without dividing by grid scale.

source

MeshArrays.gradient — Method.

gradient(inFLD::gcmfaces,iDXC::gcmfaces,iDYC::gcmfaces)

Compute spatial derivatives and multiply by inverse grid scale.

source

MeshArrays.gradient — Method.

gradient(inFLD::gcmfaces)

Compute spatial derivatives.

source

MeshArrays.mask — Method.

mask(fld::gcmfaces, val::Number)

Replace noval instances with val

source

MeshArrays.mask — Method.

mask(fld::gcmfaces, val::Number)

Replace non finite values with val

source

MeshArrays.mask — Method.

mask(fld::gcmfaces)

Call mask(fld,NaN)

source

MeshArrays.read_bin — Method.

read_bin(fil::String,kt,kk,prec::DataType)

Read model output from binary file and convert to gcmfaces structure.

source

MeshArrays.read_bin — Method.

read_bin(fil,[],[],prec)

Read model output from binary file and convert to gcmfaces structure.

source

MeshArrays.read_bin — Method.

read_bin(fil::String)

Read model output from binary file and convert to gcmfaces structure.

source

MeshArrays.smooth — Method.

smooth(FLD::gcmfaces,DXCsm::gcmfaces,DYCsm::gcmfaces)

Smooth out scales below threshold

source

MeshArrays.gcmfaces — Type.

gcmfaces{T, N}

gcmfaces data structure.

source

MeshArrays.gcmfaces — Method.

gcmfaces()

gcmfaces constructor.

source

MeshArrays.gcmfaces — Method.

gcmfaces(A::AbstractGcmfaces{T, N}) where {T,N}

gcmfaces constructor.

source

MeshArrays.gcmfaces — Method.

gcmfaces(nFaces::Int,grTopo::String,
  v1::Array{Array{T,N},1}) where {T,N}

gcmfaces constructor.

source

MeshArrays.gcmfaces — Method.

gcmfaces(nFaces::Int,grTopo::String,::Type{T},fSize::Array{NTuple{N, Int}},
  aSize::NTuple{N,Int}) where {T,N}

gcmfaces constructor.

source

MeshArrays.gcmsubset — Type.

gcmsubset{T, N}

gcmsubset data structure.

source

MeshArrays.LatCircles — Method.

LatCircles(LatValues)

Compute integration pathes along latitude circles

source

MeshArrays.TransportThrough — Method.

TransportThrough(VectorField,IntegralPath)

Computes transport through in integration path

source