MeshArrays.jl documentation

MeshArrays.jl documentation

MeshArrays.jl composite types contain array collections where arrays are typically inter-connected at their edges. The package provides exchange functions that transfer data between neighbor arrays so that the user can extend the domain of computation as needed.

Contents

MeshArrays.jl documentation

Contents

Main Features

Index

API / Functions

Main Features

To be continued...

Index

MeshArrays.gcmfaces

MeshArrays.gcmfaces

MeshArrays.gcmfaces

MeshArrays.gcmfaces

MeshArrays.gcmfaces

MeshArrays.gcmsubset

MeshArrays.GCMGridLoad

MeshArrays.GCMGridOnes

MeshArrays.GCMGridSpec

MeshArrays.convergence

MeshArrays.convert2array

MeshArrays.convert2array

MeshArrays.convert2gcmfaces

MeshArrays.convert2gcmfaces

MeshArrays.demo1

MeshArrays.demo2

MeshArrays.exchange

MeshArrays.fijind

MeshArrays.fsize

MeshArrays.fsize

MeshArrays.fsize

MeshArrays.fsize

MeshArrays.gradient

MeshArrays.gradient

MeshArrays.gradient

MeshArrays.mask

MeshArrays.mask

MeshArrays.mask

MeshArrays.read_bin

MeshArrays.read_bin

MeshArrays.read_bin

MeshArrays.smooth

API / Functions

MeshArrays.GCMGridLoad — Method.

GCMGridLoad()

Loads grid variables from files located in grDir set by GCMGridSpec.

Grid variables are XC, XG, YC, YG, RAC, RAZ, DXC, DXG, DYC, DYG, hFacC, hFacS, hFacW, Depth based on the MITgcm naming convention.

source

MeshArrays.GCMGridOnes — Method.

GCMGridOnes(grTp,nF,nP)

Define all-1 grid variables instead of using GCMGridSpec & GCMGridLoad.

source

MeshArrays.GCMGridSpec — Method.

GCMGridSpec(gridName)

Set global variables in the module scope for grDir, nFaces, grTopo, ioSize, facesSize, ioPrec using hard-coded values for LLC90, CS32, LL360 (for now).

source

MeshArrays.convergence — Method.

convergence(uFLD::gcmfaces,vFLD::gcmfaces)

Compute convergence of a vector field

source

MeshArrays.convert2array — Method.

convert2array(fld::gcmfaces)

Convert gcmfaces to array

source

MeshArrays.convert2array — Method.

convert2array(fld::Array{T,N}) where {T,N}

Convert array to gcmfaces

source

MeshArrays.convert2gcmfaces — Method.

convert2gcmfaces(fld::Array)

Convert gcmfaces to mitgcm output

source

MeshArrays.convert2gcmfaces — Method.

convert2gcmfaces(fld::gcmfaces)

Convert mitgcm output to gcmfaces

source

MeshArrays.demo1 — Method.

demo1(gridChoice)

Demo overloaded arithmetic, exchange, GCMGridLoad, gradient, etc.

source

MeshArrays.demo2 — Method.

demo2()

Demo at higher level using smooth()

source

MeshArrays.exchange — Method.

exchange(fld::gcmfaces)

Exchange / transfer data between neighbor arrays. Other methods are

exchange(fld::gcmfaces,N::Integer)
exchange(u::gcmfaces,v::gcmfaces)
exchange(u::gcmfaces,v::gcmfaces,N::Integer)

source

MeshArrays.fijind — Method.

fijind(A::gcmfaces,ij::Int)

fijind convenience function.

source

MeshArrays.fsize — Method.

fsize(A::AbstractGcmfaces{T, N}) where {T,N}

fsize convenience function.

source

MeshArrays.fsize — Method.

fsize(A::Array{Array{T,N},1}) where {T,N}

fsize convenience function.

source

MeshArrays.fsize — Method.

fsize(A::AbstractGcmfaces{T, N},i::Int) where {T,N}

fsize convenience function.

source

MeshArrays.fsize — Method.

fsize(A::Array{Array{T,N},1},i::Int) where {T,N}

fsize convenience function.

source

MeshArrays.gradient — Method.

gradient(inFLD::gcmfaces,doDIV::Bool)

Compute spatial derivatives with or without dividing by grid scale.

source

MeshArrays.gradient — Method.

gradient(inFLD::gcmfaces,iDXC::gcmfaces,iDYC::gcmfaces)

Compute spatial derivatives and multiply by inverse grid scale.

source

MeshArrays.gradient — Method.

gradient(inFLD::gcmfaces)

Compute spatial derivatives.

source

MeshArrays.mask — Method.

mask(fld::gcmfaces, val::Number)

Replace noval instances with val

source

MeshArrays.mask — Method.

mask(fld::gcmfaces, val::Number)

Replace non finite values with val

source

MeshArrays.mask — Method.

mask(fld::gcmfaces)

Call mask(fld,NaN)

source

MeshArrays.read_bin — Method.

read_bin(fil::String,kt,kk,prec::DataType)

Read model output from binary file and convert to gcmfaces structure.

source

MeshArrays.read_bin — Method.

read_bin(fil,[],[],prec)

Read model output from binary file and convert to gcmfaces structure.

source

MeshArrays.read_bin — Method.

read_bin(fil::String)

Read model output from binary file and convert to gcmfaces structure.

source

MeshArrays.smooth — Method.

smooth(FLD::gcmfaces,DXCsm::gcmfaces,DYCsm::gcmfaces)

Smooth out scales below threshold

source

MeshArrays.gcmfaces — Type.

gcmfaces{T, N}

gcmfaces data structure.

source

MeshArrays.gcmfaces — Method.

gcmfaces()

gcmfaces constructor.

source

MeshArrays.gcmfaces — Method.

gcmfaces(A::AbstractGcmfaces{T, N}) where {T,N}

gcmfaces constructor.

source

MeshArrays.gcmfaces — Method.

gcmfaces(nFaces::Int,grTopo::String,
  v1::Array{Array{T,N},1}) where {T,N}

gcmfaces constructor.

source

MeshArrays.gcmfaces — Method.

gcmfaces(nFaces::Int,grTopo::String,::Type{T},fSize::Array{NTuple{N, Int}},
  aSize::NTuple{N,Int}) where {T,N}

gcmfaces constructor.

source

MeshArrays.gcmsubset — Type.

gcmsubset{T, N}

gcmsubset data structure.

source