MeshArrays.jl documentation

MeshArrays.jl documentation

MeshArrays.jl composite types contain array collections where arrays are typically inter-connected at their edges. The package provides exchange functions that transfer data between neighbor arrays so that the user can extend the domain of computation as needed.

Contents

MeshArrays.jl documentation

Contents

Main Features

Index

API / Functions

Main Features

To be continued...

Index

MeshArrays.gcmfaces

MeshArrays.gcmfaces

MeshArrays.gcmfaces

MeshArrays.gcmfaces

MeshArrays.gcmfaces

MeshArrays.gcmsubset

MeshArrays.fijind

MeshArrays.fsize

MeshArrays.fsize

MeshArrays.fsize

API / Functions

MeshArrays.fijind — Method.

fijind(A::gcmfaces,ij::Int)

fijind convenience function.

source

MeshArrays.fsize — Method.

fsize(A::Array{Array{T,N},1}) where {T,N}

fsize convenience function.

source

MeshArrays.fsize — Method.

fsize(A::AbstractGcmfaces{T, N}) where {T,N}

fsize convenience function.

source

MeshArrays.fsize — Method.

fsize(A::Array{Array{T,N},1},i::Int) where {T,N}

fsize convenience function.

source

MeshArrays.gcmfaces — Type.

gcmfaces{T, N}

gcmfaces data structure.

source

MeshArrays.gcmfaces — Method.

gcmfaces()

gcmfaces constructor.

source

MeshArrays.gcmfaces — Method.

gcmfaces(A::AbstractGcmfaces{T, N}) where {T,N}

gcmfaces constructor.

source

MeshArrays.gcmfaces — Method.

gcmfaces(nFaces::Int,grTopo::String,
  v1::Array{Array{T,N},1}) where {T,N}

gcmfaces constructor.

source

MeshArrays.gcmfaces — Method.

gcmfaces(nFaces::Int,grTopo::String,::Type{T},
  fSize::Array{NTuple{N, Int}},
  aSize::NTuple{N,Int}) where {T,N}

gcmfaces constructor.

source

MeshArrays.gcmsubset — Type.

gcmsubset{T, N}

gcmsubset data structure.

source