API

API Guide

MeshArrays.AbstractMeshArray

MeshArrays.gcmarray

MeshArrays.gcmfaces

MeshArrays.gcmgrid

MeshArrays.gcmsubset

MeshArrays.gcmvector

Base.findall

Base.read

Base.write

MeshArrays.GridLoad

MeshArrays.GridOfOnes

MeshArrays.GridSpec

MeshArrays.LatitudeCircles

MeshArrays.MapWetPoints

MeshArrays.MaskWetPoints

MeshArrays.MatrixForPoisson

MeshArrays.ScalarPotential

MeshArrays.SeedWetPoints

MeshArrays.ThroughFlow

MeshArrays.VectorPotential

MeshArrays.convergence

MeshArrays.demo1

MeshArrays.demo2

MeshArrays.demo3

MeshArrays.exchange

MeshArrays.fijind

MeshArrays.findtiles

MeshArrays.getindexetc

MeshArrays.gradient

MeshArrays.mask

MeshArrays.nFacesEtc

MeshArrays.smooth

MeshArrays.AbstractMeshArray — Type.

AbstractMeshArray{T, N}

Subtype of AbstractArray{T, N}

source

MeshArrays.gcmgrid — Type.

gcmgrid

gcmgrid data structure. Available constructors:

gcmgrid(path::String, class::String, nFaces::Int,
        fSize::Array{NTuple{2, Int},1}, ioSize::Array{Int64,2},
        ioPrec::Type, read::Function, write::Function)

source

MeshArrays.GridLoad — Method.

GridLoad(mygrid::gcmgrid)

Return a Dict of grid variables read from files located in mygrid.path (see ?GridSpec).

Based on the MITgcm naming convention, grid variables are:

XC, XG, YC, YG, AngleCS, AngleSN, hFacC, hFacS, hFacW, Depth.

RAC, RAW, RAS, RAZ, DXC, DXG, DYC, DYG.

DRC, DRF, RC, RF (one-dimensional)

source

MeshArrays.GridOfOnes — Method.

GridOfOnes(grTp,nF,nP)

Define all-ones grid variables instead of using GridSpec & GridLoad.

source

MeshArrays.GridSpec — Function.

GridSpec(gridName)

Return a gmcgrid specification that provides grid files path, class, nFaces, ioSize, facesSize, ioPrec, & a read function (not yet) using hard-coded values for "LLC90", "CS32", "LL360" (for now).

source

MeshArrays.LatitudeCircles — Method.

LatitudeCircles(LatValues,GridVariables::Dict)

Compute integration paths that follow latitude circles

source

MeshArrays.ScalarPotential — Method.

ScalarPotential(TrspCon)

Scalar potential inversion.

TrspPot=ScalarPotential(TrspCon)

source

MeshArrays.ThroughFlow — Method.

ThroughFlow(VectorField,IntegralPath,GridVariables::Dict)

Compute transport through an integration path

source

MeshArrays.VectorPotential — Function.

VectorPotential(TrspX,TrspY,GridVariables,method::Int=1)

Vector potential inversion.

TrspPot=ScalarPotential(TrspCon)

source

MeshArrays.convergence — Method.

convergence(uFLD::MeshArray,vFLD::MeshArray)

Compute convergence of a vector field

source

MeshArrays.exchange — Method.

exchange(fld::MeshArray)

Exchange / transfer data between neighboring arrays. Other methods are

exchange(fld::MeshArray,N::Integer)
exchange(u::MeshArray,v::MeshArray)
exchange(u::MeshArray,v::MeshArray,N::Integer)

source

MeshArrays.findtiles — Method.

findtiles(ni::Int,nj::Int,mygrid::gcmgrid)
findtiles(ni::Int,nj::Int,grid::String="llc90",gridParentDir="./")

Return a MeshArray map of tile indices for tile size ni,nj

source

MeshArrays.gradient — Method.

gradient(inFLD::MeshArray,GridVariables::Dict)

Compute spatial derivatives. Other methods:

gradient(inFLD::MeshArray,GridVariables::Dict,doDIV::Bool)
gradient(inFLD::MeshArray,iDXC::MeshArray,iDYC::MeshArray)

source

MeshArrays.mask — Method.

mask(fld::MeshArray, val::Number)

Replace non finite values with val. Other methods:

mask(fld::MeshArray)
mask(fld::MeshArray, val::Number, noval::Number)

source

MeshArrays.smooth — Method.

smooth(FLD::MeshArray,DXCsm::MeshArray,DYCsm::MeshArray,GridVariables::Dict)

Smooth out scales below DXCsm / DYCsm via diffusion

source

MeshArrays.gcmarray — Type.

gcmarray{T, N}

gcmarray data structure. Available constructors:

gcmarray{T,N}(grid::gcmgrid,f::Array{Array{T,2},N},
         fSize::Array{NTuple{N, Int}},fIndex::Array{Int,1})

gcmarray(grid::gcmgrid,f::Array{Array{T,2},N}) where {T,N}
gcmarray(grid::gcmgrid,f::Array{Array{T,N},1}) where {T,N}

gcmarray(grid::gcmgrid,fSize::Array{NTuple{N, Int}},fIndex::Array{Int,1})
gcmarray(<same as above>,n3::Int)
gcmarray(<same as above>,n3::Int,n4::Int)

gcmarray(grid::gcmgrid)
gcmarray(grid::gcmgrid,::Type{T})
gcmarray(grid::gcmgrid,::Type{T},n3::Int)
gcmarray(grid::gcmgrid,::Type{T},n3::Int,n4::Int)

source

MeshArrays.gcmfaces — Type.

gcmfaces{T, N}

gcmfaces data structure. Available constructors:

gcmfaces{T,N}(grid::gcmgrid,f::Array{Array{T,N},1},
         fSize::Array{NTuple{N, Int}}, aSize::NTuple{N,Int})

gcmfaces(grid::gcmgrid,v1::Array{Array{T,N},1}) where {T,N}
gcmfaces(grid::gcmgrid,::Type{T},
         fSize::Array{NTuple{N, Int}}, aSize::NTuple{N,Int}) where {T,N}

gcmfaces(grid::gcmgrid)
gcmfaces(grid::gcmgrid,::Type{T})
gcmfaces(grid::gcmgrid,::Type{T},n3::Int)

source

MeshArrays.gcmsubset — Type.

gcmsubset{T, N}

gcmsubset data structure for subsets of gcmfaces. Available constructors:

gcmsubset{T,N}(grid::gcmgrid,f::Array{Array{T,N},1},
               fSize::Array{NTuple{N, Int}},aSize::NTuple{N, Int},
               i::Array{Array{T,N},1},iSize::Array{NTuple{N, Int}})
gcmsubset(grid::gcmgrid,::Type{T},fSize::Array{NTuple{N, Int}},
          aSize::NTuple{N,Int},dims::NTuple{N,Int}) where {T,N}

source

MeshArrays.gcmvector — Type.

gcmvector{T, N}

gcmvector data structure that can be used for subsetting and indexing into a gcmarray.

gcmvector{T,N}(grid::gcmgrid,f::Array{Array{T,1},N},
         fSize::Array{NTuple{N, Int}},fIndex::Array{Int,1})

source

Base.findall — Method.

findall(A::gcmarray{Bool})

Return a gcmvector of the true indices in A. This allows:

findall(A.<0) #gcmvector of CartesianIndex{2}
A[findall(A.<0)] #gcmvector of eltype(A)
view(A,findall(A.<0)) #CatView of eltype(A)

A[findall(A.<0)]=B[findall(A.<0)]
A[findall(A.<0)].=view(B,findall(A.<0))
A[findall(A.<0)].=NaN

source

Base.read — Method.

read(fil::String,x::MeshArray)

Read binary file to MeshArray. Other methods:

read(xx::Array,x::MeshArray) #from Array

source

Base.write — Method.

write(fil::String,x::MeshArray)

Write MeshArray to binary file. Other methods:

write(xx::Array,x::MeshArray) #to Array

source

MeshArrays.MapWetPoints — Method.

MapWetPoints(mskWet)

Mapping from global array to global ocean vector.

(Kvec,Lvec,Kmap,Lmap)=MapWetPoints(mskWet)

source

MeshArrays.MaskWetPoints — Method.

MaskWetPoints(TrspCon)

Mask land points with NaN.

(TrspCon, mskWet, mskDry)=MaskWetPoints(TrspCon)

source

MeshArrays.MatrixForPoisson — Method.

MatrixForPoisson(TrspCon,mskWet,mskDry,Kvec,Lvec,Kmap,Lmap)

Assemble sparse matrix using mskWet, Kvec, Lvec directly and Kmap, Lmap via SeedWetPoints

A=MatrixForPoisson(TrspCon,mskWet,mskDry,Kvec,Lvec,Kmap,Lmap)

source

MeshArrays.SeedWetPoints — Method.

SeedWetPoints(tmp::MeshArray,Kmap::MeshArray,Lmap::MeshArray,I...)

Seed a subset of grid points.

(FLDones,FLDkkFROM)=SeedWetPoints(tmp::MeshArray,Kmap::MeshArray,Lmap::MeshArray,I...)

source

MeshArrays.demo1 — Method.

demo1(gridChoice::String)

Demonstrate basic functionalities (load grid, arithmetic, exchange, gradient, etc.). Call sequence:

!isdir("GRID_LLC90") ? error("missing files") : nothing

(D,Dexch,Darr,DD)=MeshArrays.demo1("LLC90");

source

MeshArrays.demo2 — Method.

demo2()

Demonstrate higher level functions using smooth. Call sequence:

(Rini,Rend,DXCsm,DYCsm)=MeshArrays.demo2();

using Plots; plotlyjs()
include(joinpath(dirname(pathof(MeshArrays)),"Plots.jl"))
heatmap(Rend,title="smoothed noise",clims=(-0.5,0.5))
heatmap(Rini,title="raw noise",clims=(-0.5,0.5))

#gr(); contour(Rend,clims=(-0.5,0.5),levels=-0.5:0.1:0.5,fill=true)

source

MeshArrays.demo3 — Method.

demo3()

Demonstrate ocean transport computations. Call sequence:

(UV,LC,Tr)=MeshArrays.demo3();
using Plots; plot(Tr/1e6,title="meridional transport")

using Plots; plotlyjs()
include(joinpath(dirname(pathof(MeshArrays)),"Plots.jl"))
heatmap(UV["U"],title="U comp.",clims=(-1e7,1e7))
heatmap(UV["V"],title="V comp.",clims=(-1e7,1e7))

source

MeshArrays.fijind — Method.

fijind(A::gcmfaces,ij::Int)

Compute face and local indices (f,j,k) from global index (ij).

(needed in other types?)

source

MeshArrays.getindexetc — Method.

getindexetc(A::gcmarray, I::Vararg{_}) where {T,N}

Same as getindex but also returns the face size and index

source

MeshArrays.nFacesEtc — Method.

nFacesEtc(a::gcmarray)

Return nFaces, n3 (1 in 2D case; >1 otherwise)

source