API Guide

MeshArrays.AbstractMeshArray

MeshArrays.gcmarray

MeshArrays.gcmfaces

MeshArrays.gcmgrid

MeshArrays.gcmsubset

MeshArrays.gcmvector

MeshArrays.varmeta

Base.findall

Base.read

Base.write

MeshArrays.GridAddWS!

MeshArrays.GridLoad

MeshArrays.GridOfOnes

MeshArrays.GridSpec

MeshArrays.Interpolate

MeshArrays.InterpolationFactors

MeshArrays.LatitudeCircles

MeshArrays.MapWetPoints

MeshArrays.MaskWetPoints

MeshArrays.MatrixForPoisson

MeshArrays.ParaCoeffs

MeshArrays.PolygonAngle

MeshArrays.QuadArrays

MeshArrays.QuadCoeffs

MeshArrays.ScalarPotential

MeshArrays.SeedWetPoints

MeshArrays.StereographicProjection

MeshArrays.ThroughFlow

MeshArrays.Tiles

MeshArrays.Tiles

MeshArrays.VectorPotential

MeshArrays.convergence

MeshArrays.exchange

MeshArrays.fijind

MeshArrays.getindexetc

MeshArrays.gradient

MeshArrays.mask

MeshArrays.nFacesEtc

MeshArrays.simple_periodic_domain

MeshArrays.smooth

NearestNeighbors.knn

MeshArrays.AbstractMeshArray — Type

AbstractMeshArray{T, N}

Subtype of AbstractArray{T, N}

MeshArrays.gcmarray — Type

gcmarray{T, N, AT}

gcmarray data structure. Available constructors:

gcmarray{T,N,AT}(grid::gcmgrid,meta::varmeta,f::Array{AT,N},
         fSize::Array{NTuple{N, Int}},fIndex::Array{Int,1},v::String)

gcmarray(grid::gcmgrid,f::Array{Array{T,2},N}) where {T,N}
gcmarray(grid::gcmgrid,f::Array{Array{T,N},1}) where {T,N}

gcmarray(grid::gcmgrid,fSize::Array{NTuple{N, Int}},fIndex::Array{Int,1})
gcmarray(<same as above>,n3::Int)
gcmarray(<same as above>,n3::Int,n4::Int)

gcmarray(grid::gcmgrid)
gcmarray(grid::gcmgrid,::Type{T})
gcmarray(grid::gcmgrid,::Type{T},n3::Int)
gcmarray(grid::gcmgrid,::Type{T},n3::Int,n4::Int)

MeshArrays.gcmfaces — Type

gcmfaces{T, N}

gcmfaces data structure. Available constructors:

gcmfaces{T,N}(grid::gcmgrid,f::Array{Array{T,N},1},
         fSize::Array{NTuple{N, Int}}, aSize::NTuple{N,Int})

gcmfaces(grid::gcmgrid,v1::Array{Array{T,N},1}) where {T,N}
gcmfaces(grid::gcmgrid,::Type{T},
         fSize::Array{NTuple{N, Int}}, aSize::NTuple{N,Int}) where {T,N}

gcmfaces(grid::gcmgrid)
gcmfaces(grid::gcmgrid,::Type{T})
gcmfaces(grid::gcmgrid,::Type{T},n3::Int)

MeshArrays.gcmgrid — Type

gcmgrid

gcmgrid data structure. Available constructors:

gcmgrid(path::String, class::String,
        nFaces::Int, fSize::Array{NTuple{2, Int},1},
        ioSize::Array{Int64,2}, ioPrec::Type,
        read::Function, write::Function)

MeshArrays.gcmsubset — Type

gcmsubset{T, N}

gcmsubset data structure for subsets of gcmfaces. Available constructors:

gcmsubset{T,N}(grid::gcmgrid,f::Array{Array{T,N},1},
               fSize::Array{NTuple{N, Int}},aSize::NTuple{N, Int},
               i::Array{Array{T,N},1},iSize::Array{NTuple{N, Int}})
gcmsubset(grid::gcmgrid,::Type{T},fSize::Array{NTuple{N, Int}},
          aSize::NTuple{N,Int},dims::NTuple{N,Int}) where {T,N}

MeshArrays.gcmvector — Type

gcmvector{T, N}

gcmvector data structure that can be used for subsetting and indexing into a gcmarray.

gcmvector{T,N}(grid::gcmgrid,f::Array{Array{T,1},N},
         fSize::Array{NTuple{N, Int}},fIndex::Array{Int,1})

MeshArrays.varmeta — Type

varmeta

varmeta data structure. By default, unit is 1.0 (non-dimensional), position is fill(0.5,3) (cell center), and name / long_name is unknown.

Available constructors:

varmeta(unit::Union{Unitful.AbstractQuantity,Number},position::Array{Float64,1},
        name::String,long_name::String)

And:

defaultmeta = varmeta(1.0,fill(0.5,3),"unknown","unknown")

Base.findall — Method

findall(A::gcmarray{Bool})

Return a gcmvector of the true indices in A. This allows:

findall(A.<0) #gcmvector of CartesianIndex{2}
A[findall(A.<0)] #gcmvector of eltype(A)
view(A,findall(A.<0)) #CatView of eltype(A)

A[findall(A.<0)]=B[findall(A.<0)]
A[findall(A.<0)].=view(B,findall(A.<0))
A[findall(A.<0)].=NaN

Base.read — Method

read(fil::String,x::MeshArray)

Read binary file to MeshArray. Other methods:

read(xx::Array,x::MeshArray) #from Array

Base.write — Method

write(fil::String,x::MeshArray)

Write MeshArray to binary file. Other methods:

write(xx::Array,x::MeshArray) #to Array

MeshArrays.GridAddWS! — Method

GridAddWS!(Γ::Dict)

Compute XW, YW, XS, and YS (vector field locations) from XC, YC (tracer field locations) and add them to Γ.

Γ=GridLoad(GridSpec("LatLonCap","GRID_LLC90/"))
GridAddWS!(Γ)

MeshArrays.GridLoad — Method

GridLoad(γ::gcmgrid)

Return a Dict of grid variables read from files located in γ.path (see ?GridSpec).

Based on the MITgcm naming convention, grid variables are:

XC, XG, YC, YG, AngleCS, AngleSN, hFacC, hFacS, hFacW, Depth.

RAC, RAW, RAS, RAZ, DXC, DXG, DYC, DYG.

DRC, DRF, RC, RF (one-dimensional)

MeshArrays.GridOfOnes — Method

GridOfOnes(grTp,nF,nP)

Define all-ones grid variables instead of using GridSpec & GridLoad. E.g.

γ,Γ=GridOfOnes("CubeSphere",6,20);

MeshArrays.GridSpec — Function

GridSpec(GridName,GridParentDir="./")

Return a gmcgrid specification that provides grid files path, class, nFaces, ioSize, facesSize, ioPrec, & a read function (not yet) using hard-coded values for "PeriodicDomain", "PeriodicChannel", "CubeSphere", and `"LatLonCap" for now.

MeshArrays.Interpolate — Method

Interpolate(z_in::MeshArray,f,i,j,w)

lon=[i for i=-179.5:1.0:179.5, j=-89.5:1.0:89.5]
lat=[j for i=-179.5:1.0:179.5, j=-89.5:1.0:89.5]

Γ=GridLoad(GridSpec("LatLonCap","GRID_LLC90/"))
(f,i,j,w,j_f,j_x,j_y)=InterpolationFactors(Γ,vec(lon),vec(lat))
DD=Interpolate(Γ["Depth"],f,i,j,w)

using Plots
contourf(vec(lon[:,1]),vec(lat[1,:]),DD,clims=(0.,6000.))

MeshArrays.InterpolationFactors — Method

InterpolationFactors(Γ,lon::Array{T,1},lat::Array{T,1})

Compute interpolation coefficients etc from grid Γ to lon,lat

lon=collect(45.:0.1:46.); lat=collect(60.:0.1:61.)
(f,i,j,w,j_f,j_x,j_y)=InterpolationFactors(Γ,lon,lat)

MeshArrays.LatitudeCircles — Method

LatitudeCircles(LatValues,Γ::Dict)

Compute integration paths that follow latitude circles

MeshArrays.MapWetPoints — Method

MapWetPoints(mskWet)

Mapping from global array to global ocean vector.

(Kvec,Lvec,Kmap,Lmap)=MapWetPoints(mskWet)

MeshArrays.MaskWetPoints — Method

MaskWetPoints(TrspCon)

Mask land points with NaN.

(TrspCon, mskWet, mskDry)=MaskWetPoints(TrspCon)

MeshArrays.MatrixForPoisson — Method

MatrixForPoisson(TrspCon,mskWet,mskDry,Kvec,Lvec,Kmap,Lmap)

Assemble sparse matrix using mskWet, Kvec, Lvec directly and Kmap, Lmap via SeedWetPoints

A=MatrixForPoisson(TrspCon,mskWet,mskDry,Kvec,Lvec,Kmap,Lmap)

MeshArrays.ParaCoeffs — Function

ParaCoeffs(px,py,ox=[],oy=[])

Compute bilinear interpolation coefficients for ox,oy within px,py by remapping these parallelograms to the unit square.

px,py are Mx4 matrices where each line specifies one quadrilateral.

ox,oy are MxP position matrices

pw (output) are the MxPx4 bilinear interpolation weights

x=1.0; y=1.0 #Try send the corners to unit square corners?
println(vec(ParaCoeffs([0., 2., 3., 1.]',[0., 0., 1., 1.]',x,y)))
println(vec(QuadCoeffs([0., 2.01, 3., 1.]',[0., 0., 1., 1.]',x,y)))

MeshArrays.PolygonAngle — Function

PolygonAngle(px,py,x=[],y=[])

Compute sum of interior angles for polygons or points-to-polygons (when px,py,x,y is provided as input). px,py are MxN matrices where each line specifies one polygon. (optional) x,y are position vectors.

px=[0. 0. 1. 1.]; py=[0. 1. 1. 0.];
x=collect(-1.0:0.25:2.0); y=x;
PolygonAngle(px,py,x,y)

MeshArrays.QuadArrays — Method

QuadArrays(x_grid,y_grid)

Transform xgrid,ygrid (size ni+2,nj+2) into xquad,yquad,iquad,jquad quadrilaterals (size ni+1*nj+1,4) where iquad,jquad are point indices

MeshArrays.QuadCoeffs — Function

QuadCoeffs(px,py,ox=[],oy=[])

Compute bilinear interpolation coefficients for ox,oy within px,py by remapping these quadrilaterals to the unit square.

px,py are Mx4 matrices where each line specifies one quadrilateral.

ox,oy are MxP position matrices

ow (output) are the MxPx4 bilinear interpolation weights

QuadCoeffs([-1., 8., 13., -4.]',[-1., 3., 11., 8.]',0.,6.)
QuadCoeffs([0., 2., 3., 1.]',[0., 0., 1., 1.]',0.1,0.1)

MeshArrays.ScalarPotential — Method

ScalarPotential(TrspCon)

Scalar potential inversion.

TrspPot=ScalarPotential(TrspCon)

MeshArrays.SeedWetPoints — Method

SeedWetPoints(tmp::MeshArray,Kmap::MeshArray,Lmap::MeshArray,I...)

Seed a subset of grid points.

(FLDones,FLDkkFROM)=SeedWetPoints(tmp::MeshArray,Kmap::MeshArray,Lmap::MeshArray,I...)

MeshArrays.StereographicProjection — Method

StereographicProjection(XC0,YC0,XC,YC)

Apply stereographic projection that puts XC0,YC0 at 0.0,0.0 to target point(s) XC,YC

lon=collect(45.:0.1:46.); lat=collect(60.:0.1:61.)
x,y=StereographicProjection(45.,60.,lon,lat)

MeshArrays.ThroughFlow — Method

ThroughFlow(VectorField,IntegralPath,Γ::Dict)

Compute transport through an integration path

MeshArrays.Tiles — Method

Tiles(τ::Array{Dict},x::MeshArray)

Return an Array of tiles which cover x according to tile partition τ.

γ=GridSpec("LatLonCap","GRID_LLC90/")
d=γ.read(γ.path*"Depth.data",MeshArray(γ,γ.ioPrec))
τ=Tiles(γ,30,30)
td=Tiles(τ,d)

MeshArrays.Tiles — Method

Tiles(γ::gcmgrid,ni::Int,nj::Int)

Define sudomain tiles of size ni,nj. Each tile is defined by a Dict where tile,face,i,j correspond to tile ID, face ID, index ranges.

γ=GridSpec("LatLonCap","GRID_LLC90/")
τ=Tiles(γ,30,30)

MeshArrays.VectorPotential — Function

VectorPotential(TrspX,TrspY,Γ,method::Int=1)

Vector potential inversion.

TrspPot=ScalarPotential(TrspCon)

MeshArrays.convergence — Method

convergence(uFLD::MeshArray,vFLD::MeshArray)

Compute convergence of a vector field

MeshArrays.exchange — Method

exchange(fld::MeshArray)

Exchange / transfer data between neighboring arrays. Other methods are

exchange(fld::MeshArray,N::Integer)
exchange(u::MeshArray,v::MeshArray)
exchange(u::MeshArray,v::MeshArray,N::Integer)

MeshArrays.fijind — Method

fijind(A::gcmfaces,ij::Int)

Compute face and local indices (f,j,k) from global index (ij).

(needed in other types?)

MeshArrays.getindexetc — Method

getindexetc(A::gcmarray, I::Vararg{_}) where {T,N}

Same as getindex but also returns the face size and index

MeshArrays.gradient — Method

gradient(inFLD::MeshArray,Γ::Dict)

Compute spatial derivatives. Other methods:

gradient(inFLD::MeshArray,Γ::Dict,doDIV::Bool)
gradient(inFLD::MeshArray,iDXC::MeshArray,iDYC::MeshArray)

MeshArrays.mask — Method

mask(fld::MeshArray, val::Number)

Replace non finite values with val. Other methods:

mask(fld::MeshArray)
mask(fld::MeshArray, val::Number, noval::Number)

MeshArrays.nFacesEtc — Method

nFacesEtc(a::gcmarray)

Return nFaces, n3 (1 in 2D case; >1 otherwise)

MeshArrays.simple_periodic_domain — Function

simple_periodic_domain(np::Integer,nq=missing)

Set up a simple periodic domain of size np x nq

np=16 #domain size is np x np
Γ=simple_periodic_domain(np)

MeshArrays.smooth — Method

smooth(FLD::MeshArray,DXCsm::MeshArray,DYCsm::MeshArray,Γ::Dict)

Smooth out scales below DXCsm / DYCsm via diffusion

NearestNeighbors.knn — Method

knn(xgrid,ygrid::MeshArray,x,y::Array{T,1},k::Int)

Find k nearest neighbors to each point in x,y on xgrid,ygrid

lon=collect(0.1:0.5:2.1); lat=collect(0.1:0.5:2.1);
(f,i,j,c)=knn(Γ["XC"],Γ["YC"],lon,lat)