API documentation

1. Data Structures

MeshArrays.AbstractMeshArray — Type

AbstractMeshArray{T, N}

Subtype of AbstractArray{T, N}

MeshArrays.gcmarray — Type

gcmarray{T, N, AT}

gcmarray data structure. Available constructors:

gcmarray{T,N,AT}(grid::gcmgrid,meta::varmeta,f::Array{AT,N},
         fSize::Array{NTuple{N, Int}},fIndex::Array{Int,1},v::String)

gcmarray(grid::gcmgrid,f::Array{Array{T,2},N}) where {T,N}
gcmarray(grid::gcmgrid,f::Array{Array{T,N},1}) where {T,N}

gcmarray(grid::gcmgrid,fSize::Array{NTuple{N, Int}},fIndex::Array{Int,1})
gcmarray(<same as above>,n3::Int)
gcmarray(<same as above>,n3::Int,n4::Int)

gcmarray(grid::gcmgrid)
gcmarray(grid::gcmgrid,::Type{T})
gcmarray(grid::gcmgrid,::Type{T},n3::Int)
gcmarray(grid::gcmgrid,::Type{T},n3::Int,n4::Int)

MeshArrays.gcmvector — Type

gcmvector{T, N}

gcmvector data structure that can be used for subsetting and indexing into a gcmarray.

gcmvector{T,N}(grid::gcmgrid,f::Array{Array{T,1},N},
         fSize::Array{NTuple{N, Int}},fIndex::Array{Int,1})

MeshArrays.gcmfaces — Type

gcmfaces{T, N}

gcmfaces data structure. Available constructors:

gcmfaces{T,N}(grid::gcmgrid,f::Array{Array{T,N},1},
         fSize::Array{NTuple{N, Int}}, aSize::NTuple{N,Int})

gcmfaces(grid::gcmgrid,v1::Array{Array{T,N},1}) where {T,N}
gcmfaces(grid::gcmgrid,::Type{T},
         fSize::Array{NTuple{N, Int}}, aSize::NTuple{N,Int}) where {T,N}

gcmfaces(grid::gcmgrid)
gcmfaces(grid::gcmgrid,::Type{T})
gcmfaces(grid::gcmgrid,::Type{T},n3::Int)

MeshArrays.gcmgrid — Type

gcmgrid

gcmgrid data structure. Available constructors:

gcmgrid(path::String, class::String,
        nFaces::Int, fSize::Array{NTuple{2, Int},1},
        ioSize::Array{Int64,2}, ioPrec::Type,
        read::Function, write::Function)

MeshArrays.varmeta — Type

varmeta

varmeta data structure. By default, unit is 1.0 (non-dimensional), position is fill(0.5,3) (cell center), and name / long_name is unknown.

Available constructors:

varmeta(unit::Union{Unitful.AbstractQuantity,Number},position::Array{Float64,1},
        name::String,long_name::String)

And:

defaultmeta = varmeta(1.0,fill(0.5,3),"unknown","unknown")

2. Grids And I/O

MeshArrays.GridSpec — Function

GridSpec(GridName,GridParentDir="./")

Select one of the pre-defined grids (by GridName) and return the corresponding gmcgrid – a global grid specification which contains the grid files location (GridParentDir).

Possible choices for GridName:

"PeriodicDomain"

"PeriodicChannel"

"CubeSphere"

"LatLonCap"`

using MeshArrays
g = GridSpec()
g = GridSpec("PeriodicChannel",MeshArrays.GRID_LL360)
g = GridSpec("CubeSphere",MeshArrays.GRID_CS32)
g = GridSpec("LatLonCap",MeshArrays.GRID_LLC90)
isa(g,gcmgrid)

# output

true

MeshArrays.GridOfOnes — Function

GridOfOnes(grTp,nF,nP)

Define all-ones grid variables instead of using GridSpec & GridLoad. E.g.

γ,Γ=GridOfOnes("CubeSphere",6,20);

MeshArrays.simple_periodic_domain — Function

simple_periodic_domain(np::Integer,nq=missing)

Set up a simple periodic domain of size np x nq

using MeshArrays
np=16 #domain size is np x np
Γ=simple_periodic_domain(np)
isa(Γ["XC"],MeshArray)

# output

true

MeshArrays.GridLoad — Function

GridLoad(γ::gcmgrid)

Return a Dict of grid variables read from files located in γ.path (see ?GridSpec).

Based on the MITgcm naming convention, grid variables are:

XC, XG, YC, YG, AngleCS, AngleSN, hFacC, hFacS, hFacW, Depth.

RAC, RAW, RAS, RAZ, DXC, DXG, DYC, DYG.

DRC, DRF, RC, RF (one-dimensional)

using MeshArrays

γ = GridSpec("CubeSphere",MeshArrays.GRID_CS32)
#γ = GridSpec("LatLonCap",MeshArrays.GRID_LLC90)
#γ = GridSpec("PeriodicChannel",MeshArrays.GRID_LL360)

Γ = GridLoad(γ)

isa(Γ["XC"],MeshArray)

# output

true

Base.read — Function

read(fil::String,x::MeshArray)

Read binary file to MeshArray. Other methods:

read(xx::Array,x::MeshArray) #from Array

Base.write — Function

write(fil::String,x::MeshArray)

Write MeshArray to binary file. Other methods:

write(xx::Array,x::MeshArray) #to Array

MeshArrays.exchange — Function

exchange(fld::MeshArray)

Exchange / transfer data between neighboring arrays. Other methods are

exchange(fld::MeshArray,N::Integer)
exchange(u::MeshArray,v::MeshArray)
exchange(u::MeshArray,v::MeshArray,N::Integer)

MeshArrays.Tiles — Function

Tiles(γ::gcmgrid,ni::Int,nj::Int)

Define sudomain tiles of size ni,nj. Each tile is defined by a Dict where tile,face,i,j correspond to tile ID, face ID, index ranges.

using MeshArrays
γ=GridSpec("LatLonCap",MeshArrays.GRID_LLC90)
τ=Tiles(γ,30,30)

isa(τ[1],Dict)

# output

true

Tiles(τ::Array{Dict},x::MeshArray)

Return an Array of tiles which cover x according to tile partition τ.

using MeshArrays
γ=GridSpec("LatLonCap",MeshArrays.GRID_LLC90)
d=γ.read(γ.path*"Depth.data",MeshArray(γ,γ.ioPrec))
τ=Tiles(γ,30,30)
td=Tiles(τ,d)

isa(td[1],Array)

# output

true

3. Interpolation

NearestNeighbors.knn — Function

knn(xgrid,ygrid::MeshArray,x,y::Array{T,1},k::Int)

Find k nearest neighbors to each point in x,y on xgrid,ygrid

lon=collect(0.1:0.5:2.1); lat=collect(0.1:0.5:2.1);
(f,i,j,c)=knn(Γ["XC"],Γ["YC"],lon,lat)

MeshArrays.Interpolate — Function

Interpolate(z_in::MeshArray,f,i,j,w)

using MeshArrays
lon=[i for i=-179.5:1.0:179.5, j=-89.5:1.0:89.5]
lat=[j for i=-179.5:1.0:179.5, j=-89.5:1.0:89.5]

Γ=GridLoad(GridSpec("LatLonCap",MeshArrays.GRID_LLC90))
(f,i,j,w,j_f,j_x,j_y)=InterpolationFactors(Γ,vec(lon),vec(lat))
DD=Interpolate(Γ["Depth"],f,i,j,w)

using Plots
contourf(vec(lon[:,1]),vec(lat[1,:]),DD,clims=(0.,6000.))

MeshArrays.InterpolationFactors — Function

InterpolationFactors(Γ,lon::Array{T,1},lat::Array{T,1})

Compute interpolation coefficients etc from grid Γ to lon,lat

using MeshArrays
γ=GridSpec("CubeSphere",MeshArrays.GRID_CS32)
Γ=GridLoad(γ)
lon=collect(45.:0.1:46.); lat=collect(60.:0.1:61.)
(f,i,j,w,j_f,j_x,j_y)=InterpolationFactors(Γ,lon,lat)
YC=Interpolate(Γ["YC"],f,i,j,w)
extrema(i)==(9,10)

# output

true

MeshArrays.StereographicProjection — Function

StereographicProjection(XC0,YC0,XC,YC)

Apply stereographic projection that puts XC0,YC0 at 0.0,0.0 to target point(s) XC,YC

lon=collect(45.:0.1:46.); lat=collect(60.:0.1:61.)
x,y=StereographicProjection(45.,60.,lon,lat)

4. Vector Fields

MeshArrays.gradient — Function

gradient(inFLD::MeshArray,Γ::Dict)

Compute spatial derivatives. Other methods:

gradient(inFLD::MeshArray,Γ::Dict,doDIV::Bool)
gradient(inFLD::MeshArray,iDXC::MeshArray,iDYC::MeshArray)

MeshArrays.convergence — Function

convergence(uFLD::MeshArray,vFLD::MeshArray)

Compute convergence of a vector field

MeshArrays.smooth — Function

smooth(FLD::MeshArray,DXCsm::MeshArray,DYCsm::MeshArray,Γ::Dict)

Smooth out scales below DXCsm / DYCsm via diffusion

MeshArrays.ScalarPotential — Function

ScalarPotential(TrspCon)

Scalar potential inversion.

TrspPot=ScalarPotential(TrspCon)

MeshArrays.VectorPotential — Function

VectorPotential(TrspX,TrspY,Γ,method::Int=1)

Vector potential inversion.

TrspPot=ScalarPotential(TrspCon)

MeshArrays.LatitudeCircles — Function

LatitudeCircles(LatValues,Γ::Dict)

Compute integration paths that follow latitude circles

MeshArrays.ThroughFlow — Function

ThroughFlow(VectorField,IntegralPath,Γ::Dict)

Compute transport through an integration path

5. Various

MeshArrays.getindexetc — Function

getindexetc(A::gcmarray, I::Vararg{_}) where {T,N}

Same as getindex but also returns the face size and index

MeshArrays.nFacesEtc — Function

nFacesEtc(a::gcmarray)

Return nFaces, n3 (1 in 2D case; >1 otherwise)

Base.findall — Function

findall(A::gcmarray{Bool})

Return a gcmvector of the true indices in A. This allows:

findall(A.<0) #gcmvector of CartesianIndex{2}
A[findall(A.<0)] #gcmvector of eltype(A)
view(A,findall(A.<0)) #CatView of eltype(A)

A[findall(A.<0)]=B[findall(A.<0)]
A[findall(A.<0)].=view(B,findall(A.<0))
A[findall(A.<0)].=NaN

MeshArrays.mask — Function

mask(fld::MeshArray, val::Number)

Replace non finite values with val. Other methods:

mask(fld::MeshArray)
mask(fld::MeshArray, val::Number, noval::Number)